Refactor
Move tile form from CLandscape to CView
Move CTilePreview from CLandscape to CView
Improve tile picking
